Starting to move ConnectionManager and ConnectionRegistry to the Datasource namespace
