+++ src/Datasource/Error/MissingDatasourceException.php
@@ -1,5 +1,7 @@
 <?php
 /**
+ * MissingDatasourceException class
+ *
  * PHP 5
  *
  * CakePHP(tm) : Rapid Development Framework (http://cakephp.org)
@@ -13,13 +15,16 @@
  * @since         CakePHP(tm) v 3.0
  * @license       MIT License (http://www.opensource.org/licenses/mit-license.php)
  */
-namespace Cake\Error;
+namespace Cake\Datasource\Error;
+
+use Cake\Error\Exception;
 
 /**
- * Exception class to be thrown when a datasource configuration is not found
+ * Used when a datasource cannot be found.
+ *
  */
-class MissingDatasourceConfigException extends Exception {
+class MissingDatasourceException extends Exception {
 
-	protected $_messageTemplate = 'The datasource configuration "%s" was not found.';
+	protected $_messageTemplate = 'Datasource class %s could not be found. %s';
 
 }
