+++ src/TestSuite/Fixture/FixtureManager.php
@@ -19,8 +19,7 @@ namespace Cake\TestSuite\Fixture;
 use Cake\Core\App;
 use Cake\Core\Configure;
 use Cake\Core\Plugin;
-use Cake\Database\Connection;
-use Cake\Database\ConnectionManager;
+use Cake\DataSource\ConnectionManager;
 use Cake\Error;
 use Cake\TestSuite\Fixture\TestFixture;
 use Cake\TestSuite\TestCase;
