+++ tests/init.php
@@ -13,7 +13,7 @@
 
 use Cake\Cache\Cache;
 use Cake\Core\Configure;
-use Cake\Database\ConnectionManager;
+use Cake\Datasource\ConnectionManager;
 use Cake\I18n\I18n;
 use Cake\Log\Log;
 
@@ -98,7 +98,8 @@ Cache::config([
 ]);
 
 ConnectionManager::config('test', [
-	'className' => getenv('db_class'),
+	'className' => 'Cake\Database\Connection',
+	'driver' => getenv('db_class'),
 	'dsn' => getenv('db_dsn'),
 	'database' => getenv('db_database'),
 	'login' => getenv('db_login'),
